openaidocker/open1/requirements.txt
openai
python-dotenv
gradio
